Error Handling aon the POST form and UI design on the website pages

+++ virtualTravel/views.py
@@ -20,10 +20,10 @@ import re
 import json
 from virtualTravel.yelp import *
 
-
 import urllib.request
 import time
 
+MAX_UPLOAD_SIZE = 5000000
 
 def homepage(request):
     return redirect(reverse('homepage'))
@@ -37,8 +37,15 @@ def login_action(request):
         context['form'] = LoginForm()
         return render(request, 'virtualTravel/login.html', context)
 
-    form = LoginForm(request.POST)
-    context['form'] = form
+    # Check post error
+    if not 'username' in request.POST:
+        context['name_error'] = "Miss input name"
+        return render(request, 'virtualTravel/login.html', context)
+
+    if not 'password' in request.POST:
+        context['password_error'] = "Miss password"
+        return render(request, 'virtualTravel/login.html', context)
+
 
     # Creates a bound form from the request POST parameters and makes the 
     info = {
@@ -49,8 +56,11 @@ def login_action(request):
     # If no username or no password, set error message
     if request.POST['username'] == "":
         context['name_error'] = "Miss input name"
+        return render(request, 'virtualTravel/login.html', context)
+
     if request.POST['password'] == "":
         context['password_error'] = "Miss password"
+        return render(request, 'virtualTravel/login.html', context)
         
     form = LoginForm(info)
 
@@ -58,7 +68,6 @@ def login_action(request):
 
     # Validates the form.
     if not form.is_valid():
-        print("not valid")
         return render(request, 'virtualTravel/login.html', context)
 
     new_user = authenticate(username=form.cleaned_data['username'],
@@ -82,6 +91,16 @@ def register_action(request):
     # Creates a bound form from the request POST parameters and makes the 
     # form available in the request context dictionary.
 
+    list = {'username', 'password', 'confirm_password', 'email'}
+
+    for field in list:
+        if field not in request.POST:
+            print(field)
+            error = field+'_error'
+            print(error)
+            context[error] = "Miss" + " " + field
+            return render(request, 'virtualTravel/register.html', context)
+
     # Get registration from information
     info = {
         'username': request.POST['username'],
@@ -92,7 +111,7 @@ def register_action(request):
 
     # Set the error message if mal input
     if (request.POST['username'] == ""):
-        context['name_error'] = "Miss input name"
+        context['username_error'] = "Miss username"
 
     if (request.POST['password'] == ""):
         context['password_error'] = "Miss password"
@@ -115,7 +134,6 @@ def register_action(request):
 
     # Validates the form.
     if not form.is_valid():
-        print("invalid form")
         return render(request, 'virtualTravel/register.html', context)
 
     # At this point, the form data is valid.  Register and login the user.
@@ -153,20 +171,20 @@ def travel_select_action(request):
     return render(request, 'virtualTravel/travel_select.html', context)
 
 @login_required
-def travel_action(request):
-    context = {}
-
+def travel_action(request,context={}):
     profile = request.user.profile
     cities = list(profile.city_pool.all())
     errorlist = []
 
     if request.method  == 'GET':
-        context['current_page'] = 'list-City-list'  
+        if "current_page" not in context:
+            context['current_page'] = 'list-City-list'  
 
     # Create a 'Travel' if this page is entered by clicking 'Start Journey'
     if request.method == 'POST':
-        if 'departure' in request.POST:
-            context['current_page'] = 'list-City-list'
+        if ('departure' in request.POST) and ('destination' in request.POST):
+            if "current_page" not in context:
+                context['current_page'] = 'list-City-list'
             departure = request.POST['departure']
             destination = request.POST['destination']
 
@@ -174,7 +192,8 @@ def travel_action(request):
             if len(profile.current_travel.all()) == 0:
                 # Error Handling
                 if departure == 'Choose...' or destination == 'Choose...':
-                    errorlist.append('Please Select Cities for Departure and Destination')
+                    errorlist.append('Please Select Departure and Destination.')
+                
                 # Generating routes for the travel
                 route = []
                 departure_id = -1
@@ -195,23 +214,26 @@ def travel_action(request):
                 else:
                     destination_id = departure_id
 
-                # Randomly generate intermediate stops from the remaining city pool
-                num_of_stops = randint(2, 4) # 2 - 4 stops
-                if num_of_stops > len(cities): # corner case
-                    num_of_stops = len(cities)
-                stops = sample(cities, num_of_stops)
+                if departure_id != -1 and destination_id != -1:
+                    # Randomly generate intermediate stops from the remaining city pool
+                    num_of_stops = randint(1, 3) # 1 - 3 stops
+                    if num_of_stops > len(cities): # corner case
+                        num_of_stops = len(cities)
+                    stops = sample(cities, num_of_stops)
 
-                # Transfer the route into String format
-                route_str = str(departure_id) + ' '
-                for city in stops:
-                    route_str += str(city.id) + ' '
-                route_str += str(destination_id)
+                    # Transfer the route into String format
+                    route_str = str(departure_id) + ' '
+                    for city in stops:
+                        route_str += str(city.id) + ' '
+                    route_str += str(destination_id)
 
-                # Create the 'current travel' object from the generated route
-                if len(errorlist) == 0: 
-                    current_travel = Travel(num_of_stops=len(route_str.split(' ')), current_stop=0, route=route_str,
-                                             date=timezone.now(), made_by=profile, current_user=profile)
-                    current_travel.save()
+                    # Create the 'current travel' object from the generated route
+                    if len(errorlist) == 0: 
+                        current_travel = Travel(num_of_stops=len(route_str.split(' ')), current_stop=0, route=route_str,
+                                                 date=timezone.now(), made_by=profile, current_user=profile)
+                        current_travel.save()
+                elif len(errorlist) == 0:
+                    errorlist.append('Please Select a Valid City.')
     
     # rendering city content and quiz no matter GET or POST        
     if len(profile.current_travel.all()) == 0:
@@ -253,17 +275,12 @@ def travel_action(request):
     current_quiz = quiz_set[randint(0,max(0,len(quiz_set))-1)]
     context['quiz'] = current_quiz
     if profile in current_quiz.quiz_users.all():
-        print("quiz already answered by this user")
         context['quiz_answered'] = "true"
 
 
     current_city_sites=Site.objects.filter(city=current_city)
     context['current_city_sites']=current_city_sites
-    print("current_city_sites are: ")
-    for site in current_city_sites:
-        print(site.name+" id: "+(str)(site.id))
     first_site_id = current_city_sites[0].id
-    print("first site id is "+(str)(first_site_id))
     context['first_site_id']=first_site_id
     
     return render(request, 'virtualTravel/traveling.html',context)
@@ -274,15 +291,12 @@ def next_city_action(request):
     current_travel = profile.current_travel.all()[0]
     current_travel.current_stop += 1
     current_travel.save()
-    print("move to next city")
     return redirect(reverse('travel'))
 
 @login_required
 def profile_action(request):
     context = {}
-
     profile = request.user.profile
-
     cities = list(profile.city_collection.all())
 
     if len(cities) >= 0:
@@ -300,13 +314,18 @@ def profile_action(request):
 
     # Update the user picture and bio
     if request.method == 'POST':
-        profile.bio = request.POST['bio']
-        form = ProfileForm(request.POST, request.FILES, instance=profile)
+        if 'bio' in request.POST:
+            profile.bio = request.POST['bio']
         if 'picture' in request.FILES:
-            profile.picture.delete()
             picture = request.FILES['picture']
-            profile.content_type = picture.content_type
-            profile.picture.save(picture.name, picture)
+            if not picture.content_type or not picture.content_type.startswith('image'):
+                context['error'] = 'File type is not image'
+            elif picture.size > MAX_UPLOAD_SIZE:
+                context['error'] = 'File too big (max size is {0} bytes)'.format(MAX_UPLOAD_SIZE)
+            else:
+                profile.picture.delete()
+                profile.content_type = picture.content_type
+                profile.picture.save(picture.name, picture)
         profile.save()
 
     context['profile'] = profile
@@ -353,6 +372,7 @@ def searchCity_action(request):
 
 # Buy cities
 @login_required
+@transaction.atomic
 def buyCity_action(request):
      context = {}
      myProfile = request.user.profile
@@ -370,13 +390,15 @@ def buyCity_action(request):
      if 'city_name' in request.POST:
         try:
             city = City.objects.get(name = request.POST['city_name'])
-            if myProfile.gold < city.price:
-                pass
-            else:
-                # Compute the balance if user buy the city
-                goldChange(myProfile, -city.price)
-                myProfile.city_pool.add(city)
-                myProfile.save()
+
+            if city not in myProfile.city_pool.all():
+                if myProfile.gold < city.price:
+                    pass
+                else:
+                    # Compute the balance if user buy the city
+                    goldChange(myProfile, -city.price)
+                    myProfile.city_pool.add(city)
+                    myProfile.save()
         except:
             pass
 
@@ -445,7 +467,10 @@ def get_photo(request):
 def routeParser(travel):
     route = []
     for city_id in travel.route.split(' '):
-        city = City.objects.get(id=int(city_id))
+        try:
+            city = City.objects.get(id=int(city_id))
+        except:
+            city = City.objects.get(id=1)
         route.append(city)
     return route
 
@@ -456,10 +481,6 @@ def user_upload(request):
     # Check user city pool
     user_profile=request.user.profile
     user_city_pool=user_profile.city_pool
-    print("user city pool is: ")
-    for city in user_city_pool.all():
-        print(city.name)
-    print(" ")
     cities = list(user_city_pool.all())
     context = {'city_pool':cities}
     
@@ -501,6 +522,23 @@ def user_upload(request):
             context['quiz_message']="Please select a valid city."
             return render(request, 'virtualTravel/user_upload.html', context)
 
+        # check whether all parameters exist
+        if not "quiz_text" in request.POST:
+            context['quiz_message']='Missing quiz question!'
+            return render(request, 'virtualTravel/user_upload.html', context)
+        if not "quiz_option_1" in request.POST:
+            context['quiz_message']='Missing quiz option 1!'
+            return render(request, 'virtualTravel/user_upload.html', context)
+        if not "quiz_option_2" in request.POST:
+            context['quiz_message']='Missing quiz option 2!'
+            return render(request, 'virtualTravel/user_upload.html', context)
+        if not "quiz_option_3" in request.POST:
+            context['quiz_message']='Missing quiz option 3!'
+            return render(request, 'virtualTravel/user_upload.html', context)
+        if not "quiz_option_4" in request.POST:
+            context['quiz_message']='Missing quiz option 4!'
+            return render(request, 'virtualTravel/user_upload.html', context)
+
         # Get the input data and save the quiz in database
         new_quiz = Quiz(quiz_city=find_city,
                         quiz_text=request.POST['quiz_text'], 
@@ -519,7 +557,18 @@ def user_upload(request):
 
 
     # ************************** Upload site ******************************** #
-    if 'city_name' in request.POST:
+    elif 'city_name' not in request.POST:
+        context['site_message']="Missing city name!"
+    else:
+        if 'name' not in request.POST:
+            context['site_message']="Missing site name!"
+            return render(request, 'virtualTravel/user_upload.html', context)
+        if 'description' not in request.POST:
+            context['site_message']="Missing site description!"
+            return render(request, 'virtualTravel/user_upload.html', context)
+        if 'site_picture_url' not in request.POST:
+            context['site_message']="Missing site picture url!"
+            return render(request, 'virtualTravel/user_upload.html', context)
         upload_site_name=request.POST['name'].strip()
         upload_site_city=request.POST['city_name'].strip()
 
@@ -563,7 +612,7 @@ def user_upload(request):
 # check upload url is correct
 def checkurl(url):
     # Try to open the url
-    if re.match(r'(((http|ftp|https)://)?)',url, re.M|re.I):
+    if not re.match(r'((http|https):\/\/([\w.]+\/?)\S*/)',url, re.M|re.I):
         return False
     opener = urllib.request.build_opener()
     opener.addheaders = [('User-agent', 'Mozilla/49.0.2')]
@@ -581,14 +630,9 @@ def checkurl(url):
 @login_required
 def foodPage_action(request):
     if request.method == 'GET':
-        context = {}
-        user_profile = request.user.profile
-        current_travel = user_profile.current_travel.all()[0]
-        routes = current_travel.route.split(" ")
-        current_city_id = int(routes[current_travel.current_stop])
-        current_city = City.objects.get(id=current_city_id)
-        context['current_city'] = current_city
-        return render(request, 'virtualTravel/food.html', context)
+        context['current_page'] = 'list-Food-list'
+        return travel_action(request, context)
+
 
 # Search the food
 @login_required
@@ -605,7 +649,8 @@ def food_search_action(request):
         request.session['search_results'] = query_api(filters, request.session['location_input'])
         return redirect(reverse('food_result'))
     else:
-        return redirect(reverse('travel'))
+        context['current_page'] = 'list-Food-list'
+        return travel_action(request, context)
 
 # Show the food search result        
 @login_required
@@ -614,16 +659,10 @@ def food_result_action(request):
     context = {
         'places_list': search_results
     }
-    user_profile = request.user.profile
-    current_travel = user_profile.current_travel.all()[0]
-    routes = current_travel.route.split(" ")
-    current_city_id = int(routes[current_travel.current_stop])
-    current_city = City.objects.get(id=current_city_id)
-    context['current_city'] = current_city
     context['current_page'] = 'list-Food-list'
+    return travel_action(request, context)
 
-    return render(request, 'virtualTravel/traveling.html', context)
-
+@transaction.atomic
 def check_quiz(request,id,choice):
     profile = request.user.profile
     current_quiz = Quiz.objects.get(id=id)
@@ -637,7 +676,7 @@ def check_quiz(request,id,choice):
             profile.save()
             current_travel.save()
             current_quiz.quiz_users.add(profile)
-            gold = " You've been awarded 100 gold"
+            gold = " You've been awarded 100 gold."
         response = '<div class="alert alert-success" role="alert"> \
                     <button type="button" class="close" data-dismiss="alert">&times;</button>\
                     <strong>Well done!</strong> The answer is Correct!' + gold +'</div>'
